genai-function-calling: refactors to keep mcp related code in a single file
