+++ genai-function-calling/openai-agents/Dockerfile
@@ -7,7 +7,6 @@ COPY requirements.txt /tmp
 RUN --mount=type=cache,target=/root/.cache/pip pip install -r /tmp/requirements.txt
 RUN --mount=type=cache,target=/root/.cache/pip edot-bootstrap --action=install
 
-COPY main.py /
-COPY mcp_server.py /
+COPY *.py /
 
 ENTRYPOINT [ "opentelemetry-instrument", "python", "main.py" ]
