+++ genai-function-calling/openai-agents/main.py
@@ -1,6 +1,6 @@
-import argparse
 import asyncio
 import os
+import sys
 
 from httpx import AsyncClient
 from agents import (
@@ -10,11 +10,12 @@ from agents import (
     RunConfig,
     Runner,
     function_tool,
+    Tool,
 )
 from agents.tracing import GLOBAL_TRACE_PROVIDER
 from openai import AsyncAzureOpenAI
 
-from mcp_server import mcp_client_main, SERVER_ARG
+from main_mcp import run_main as mcp_main
 
 # Shut down the global tracer as it sends to the OpenAI "/traces/ingest"
 # endpoint, which we aren't using and doesn't exist on alternative backends
@@ -51,7 +52,7 @@ async def get_latest_elasticsearch_version(major_version: int = 0) -> str:
     return max(versions, key=lambda v: tuple(map(int, v.split("."))))
 
 
-async def run_agent(**agent_kwargs: dict):
+async def run_agent(tools: list[Tool]):
     model_name = os.getenv("CHAT_MODEL", "gpt-4o-mini")
     openai_client = AsyncAzureOpenAI() if os.getenv("AZURE_OPENAI_API_KEY") else None
     model = OpenAIProvider(openai_client=openai_client, use_responses=False).get_model(model_name)
@@ -59,7 +60,7 @@ async def run_agent(**agent_kwargs: dict):
         name="version_assistant",
         model=model,
         model_settings=ModelSettings(temperature=0),
-        **agent_kwargs,
+        tools=tools,
     )
 
     result = await Runner.run(
@@ -71,27 +72,13 @@ async def run_agent(**agent_kwargs: dict):
 
 
 async def main():
-    parser = argparse.ArgumentParser(
-        prog="genai-function-calling",
-        description="Fetches the latest version of Elasticsearch 8",
-    )
-    parser.add_argument(
-        "--mcp",
-        action="store_true",
-        help="Run tools via a MCP server instead of directly",
-    )
-    parser.add_argument(
-        SERVER_ARG,
-        action="store_true",
-        help="Run the MCP server",
-    )
+    """Run tools with the agent directly unless in MCP mode"""
 
-    args, _ = parser.parse_known_args()
-
-    if args.mcp:
-        await mcp_client_main(run_agent, [get_latest_elasticsearch_version], args.mcp_server)
-    else:
-        await run_agent(tools=[function_tool(strict_mode=False)(get_latest_elasticsearch_version)])
+    fns = [get_latest_elasticsearch_version]
+    if any(arg.startswith("--mcp") for arg in sys.argv):
+        await mcp_main(fns, run_agent)  # start an MCP server and use its tools
+    else:  # run the tools directly
+        await run_agent([function_tool(strict_mode=False)(fn) for fn in fns])
 
 
 if __name__ == "__main__":
